requirements_flask.txt
Flask==2.3.3
kafka-python==2.0.2
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
blinker==1.6.3
